Add publishing Physical Host
This publisher creates a structure for Physical Host in Dobby. It creates
the following identifiers:
* Physical Host
* Patch Panel of PH
* Physical Ports of PH

The publisher connects all the above using ‘part_of’ links. If the
Cluster Patch Panel doesn’t exists it is created by the publisher.
All the identifiers created by the publisher are prefixed with the
Physical Host.